Simple properties are less likely to raise errors
Conventions:
- use `dict.get()` rather than `dict[]`
- use `(foo or {})` when accessing nested dicts
- check for `None` before using string methods

These apply specifically for simple plain-python-object access of fields
and derivative fields. Methods that take external actions can and
should raise errors when preconditions are not met.

Relationale: Property accessors may be used in debugging, in string
reprs, and in other scenarios where exceptions are not expected and may
be obstacles in diagnosing other issues. Where reasonable, property
accessors should be exception-free.